mpihandles: initial cut of mapping procs to MCW ranks
This took an unexpectedly large amount of code to implement.  :-(

Hang OPAL process names off the ompi_mpidbg_comm_handle_t, and use
those to look up their rank in MCW.  A global table is filled in
during init_per_process() that maps name OPAL process names of the
procs on ompi_mpi_comm_world to MCW ranks.  If OPAL process names are
found in the future that are not in this table, they are added to the
table and given a unique (negative) rank/ID.